+++ app/components/SideNav.vue
@@ -75,7 +75,6 @@ ul {
     img {
       width: 18px;
       height: 18px;
-      filter: var(--icon-filter);
     }
   }